requirements.txt
asgiref
Django
django-smart-selects
Pillow
pytz
six
sqlparse
python-decouple
gunicorn
psycopg2-binary